Displaying item moved to new component
